refactor[#425] : useResearchMarkingList 삭제
