+++ src/features/map/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from "./useCurrentLocation";
-export * from "./useResearchMarkingList";
 export * from "./useGetMapCurrentBounds";
 export * from "./useMapParams";
